Remove legacy Account Console tests
