+++ .github/CODEOWNERS
@@ -47,4 +47,3 @@
 /js/                                                        @keycloak/ui-maintainers
 /adapters/oidc/js/                                          @keycloak/ui-maintainers
 /rest/admin-ui-ext/                                         @keycloak/ui-maintainers
-/testsuite/integration-arquillian/tests/other/base-ui/      @keycloak/ui-maintainers
