+++ testsuite/integration-arquillian/HOW-TO-RUN.md
@@ -352,17 +352,6 @@ mvn -f testsuite/integration-arquillian/tests/other/springboot-tests/pom.xml \
     [-Pspringboot27]
 ```
 
-## Base UI tests
-Similarly to Admin Console tests, these tests are focused on UI, specifically on the parts of the server that are accessed by an end user (like Login page, or Account Console).
-They are designed to work with mobile browsers (alongside the standard desktop browsers). For details on the supported browsers and their configuration please refer to [Different Browsers chapter](#different-browsers).
-#### Execution example
-```
-mvn -f testsuite/integration-arquillian/tests/other/base-ui/pom.xml \
-    clean test \
-    -Pandroid \
-    -Dappium.avd=Nexus_5X_API_27
-```
-
 ## Disabling features
 Some features in Keycloak can be disabled. To run the testsuite with a specific feature disabled use the `auth.server.feature` system property. For example to run the tests with authorization disabled run:
 ```
@@ -454,22 +443,16 @@ You can use many different real-world browsers to run the integration tests.
 Although technically they can be run with almost every test in the testsuite, they can fail with some of them as the tests often require specific optimizations for given browser. Therefore, only some of the test modules have support to be run with specific browsers.
 
 #### Mozilla Firefox
-* **Supported test modules:** `console`, `base-ui`
+* **Supported test modules:** `console`
 * **Supported version:** latest stable
 * **Driver download required:** [GeckoDriver](https://github.com/mozilla/geckodriver/releases)
 * **Run with:** `-Dbrowser=firefox -Dwebdriver.gecko.driver=path/to/geckodriver`; optionally you can specify `-Dfirefox_binary=path/to/firefox/binary`
 
 #### Google Chrome
-* **Supported test modules:** `console`, `base-ui`
+* **Supported test modules:** `console`
 * **Supported version:** latest stable
 * **Driver download required:** [ChromeDriver](https://sites.google.com/a/chromium.org/chromedriver/) that corresponds with your version of the browser
 * **Run with:** `-Dbrowser=chrome -Dwebdriver.chrome.driver=path/to/chromedriver`
-
-#### Apple Safari
-* **Supported test modules:** `base-ui`
-* **Supported version:** latest stable
-* **Driver download required:** no (the driver is bundled with macOS)
-* **Run with:** `-Dbrowser=safari`
 
 #### [DEPRECATED] Mozilla Firefox with legacy driver
 * **Supported test modules:** `console`
@@ -490,24 +473,10 @@ The tests will try to start the Appium server automatically but you can do it ma
 
 To use a mobile browser you need to create a virtual device. The most convenient way to do so is to install the desired platform's IDE - either [Android Studio](https://developer.android.com/studio/) (for Android devices) or [Xcode](https://developer.apple.com/xcode/) (for iOS devices) - then you can create a device (smartphone/tablet) there. For details please refer to documentation of those IDEs.
 
-#### Google Chrome on Android
-* **Supported test modules:** `base-ui`
-* **Supported host OS:** Windows, Linux, macOS
-* **Supported browser version:** latest stable
-* **Supported mobile OS version:** Android 7.x, 8.x
-* **Run with:** `mvn clean test -Pandroid -Dappium.avd=name_of_the_AVD` where AVD is the name of your Android Virtual Device (e.g. `Nexus_5X_API_27`)
-
 **Tips & tricks:**
 * If the AVD name contains any spaces, you need to replace them with underscores when specifying the `-Dappium.avd=...`.
 * It's probable that a freshly created device will contain an outdated Chrome version. To update to the latest version (without using the Play Store) you need to download an `.apk` for Chrome and install it with `adb install -r path/to/chrome.apk`.
 * Chrome on Android uses ChromeDriver similarly to regular desktop Chrome. The ChromeDriver is bundled with the Appium server. To use a newer ChromeDriver please follow the [Appium documentation](http://appium.io/docs/en/writing-running-appium/web/chromedriver/).
-
-#### Apple Safari on iOS
-* **Supported test modules:** `base-ui`
-* **Supported host OS:** macOS
-* **Supported browser version:** _depends on the mobile OS version_
-* **Supported mobile OS version:** iOS 11.x
-* **Run with:** `mvn clean test -Pios -Dappium.deviceName=device_name` where the device name is your device identification (e.g. `iPhone X`)
 
 ## Disabling TLS (SSL) in the tests
