--- testsuite/integration-arquillian/tests/other/base-ui/src/main/java/org/keycloak/testsuite/ui/account3/page/fragment/AbstractFragmentWithMobileLayout.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.keycloak.testsuite.ui.account3.page.fragment;
-
-import org.jboss.arquillian.drone.api.annotation.Drone;
-import org.openqa.selenium.WebDriver;
-
-/**
- * @author Vaclav Muzikar <vmuzikar@redhat.com>
- */
-public abstract class AbstractFragmentWithMobileLayout {
-    @Drone
-    private WebDriver driver;
-
-    private Boolean mobileLayout = null;
-
-    abstract protected int getMobileWidth();
-
-    public boolean isMobileLayout() {
-        if (mobileLayout == null) {
-            mobileLayout = driver.manage().window().getSize().width <= getMobileWidth();
-        }
-        return mobileLayout;
-    }
-}
