--- testsuite/integration-arquillian/tests/other/base-ui/src/main/resources/themes/localized-theme-preview/account/messages/messages_en.properties
@@ -1,2 +1,0 @@
-locale_test=Přísný jazyk
-client_localized-client=Přespříliš lokalizovaný klient
