--- testsuite/integration-arquillian/tests/other/base-ui/src/main/resources/themes/localized-theme-preview/account/messages/messages_test.properties
@@ -1,3 +1,0 @@
-locale_test=Přísný jazyk
-accountManagementWelcomeMessage=Vítejte v Keycloaku
-personalInfoHtmlTitle=Osobní údaje
