--- testsuite/integration-arquillian/tests/other/base-ui/src/main/resources/themes/localized-theme/email/messages/messages_en.properties
@@ -1,1 +1,0 @@
-locale_test=Přísný jazyk
