--- testsuite/integration-arquillian/tests/other/base-ui/src/main/resources/themes/localized-theme/login/messages/messages_test.properties
@@ -1,11 +1,0 @@
-locale_test=Přísný jazyk
-termsText=[TEST LOCALE] souhlas s podmínkami
-notMatchPasswordMessage=[TEST LOCALE] hesla se neshodují
-firstName=[TEST LOCALE] křestní jméno
-updateProfileMessage=[TEST LOCALE] aktualizovat profil
-verifyEmailMessage=[TEST LOCALE] je třeba ověřit emailovou adresu
-invalidTotpMessage=[TEST LOCALE] vložen chybný kód
-oauthGrantTitle=[TEST LOCALE] Udělit přístup {0}
-rememberMe=[TEST LOCALE] Zapamatuj si mě
-invalidUserMessage=[TEST LOCALE] Chybné jméno nebo heslo
-emailForgotTitle=[TEST LOCALE] Zapomenuté heslo
