--- testsuite/integration-arquillian/tests/other/base-ui/src/main/resources/themes/localized-theme/login/theme.properties
@@ -1,2 +1,0 @@
-parent=${theme-default-name}
-locales=en,de,lang01,lang02,lang03,lang04,lang05,test,lang06,lang07,lang08,lang09,lang10
