+++ testsuite/integration-arquillian/tests/other/webauthn/src/main/java/org/keycloak/testsuite/webauthn/pages/fragments/Sidebar.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.testsuite.ui.account3.page.fragment;
+package org.keycloak.testsuite.webauthn.pages.fragments;
 
 import org.jboss.arquillian.graphene.fragment.Root;
 import org.openqa.selenium.By;
